refactor(adapter-ui): address review — extract ai-traces client + defensive guards
coderabbit + gemini + claude review on #563:
- Extract AITrace* types + fetchAITraces into lib/ai-traces-client.ts (api.ts was
  far over the 500-line cap); export getAuthHeaders/handleUnauthorized for reuse.
- fetchAITraces: a 403 with an empty/non-JSON body now maps to `denied` (not
  `error`) so the permission message still shows behind a proxy/gateway (gemini).
  Drop the dead `res.status === 401` from the denied arm — 401 is handled upstream
  by handleUnauthorized (token clear + /login redirect), never reaches here (claude).
- Page guards: invalid/missing startedAt no longer throws RangeError (renders "—");
  endedAt uses `== null`; tokens use `?? 0`; the stale-response drop also bails on
  `signal?.aborted` to avoid setState-after-unmount (gemini).
- Tests: + query-string contract (limit/status) and a non-JSON 403 → denied case.
